@@ -1,13 +1,8 @@
 <?php
-/**
- * Created by IntelliJ IDEA.
- * User: stopka
- * Date: 13.10.17
- * Time: 12:19
- */
+
+declare(strict_types=1);
 
 namespace Stopka\OpenviduPhpClient;
-
 
 use Stopka\OpenviduPhpClient\Recording\Recording;
 use Stopka\OpenviduPhpClient\Recording\RecordingLayoutEnum;
@@ -21,41 +16,20 @@ use Stopka\OpenviduPhpClient\Session\SessionProperties;
 
 class OpenVidu
 {
-    protected const HTTP_USERNAME = "OPENVIDUAPP";
-
-    const SESSIONS_URL = "api/sessions";
-    const TOKENS_URL = "api/tokens";
-    const RECORDINGS_URL = "api/recordings";
-    const RECORDINGS_START_URL = "api/recordings/start";
-    const RECORDINGS_STOP_URL = "api/recordings/stop";
-
-    /** @var  string */
-    protected $urlOpenViduServer;
-
-    /** @var string */
-    protected $secret;
-
-    /** @var bool */
-    protected $sslCheck = true;
 
     /** @var RestClient */
-    protected $restClient;
+    protected RestClient $restClient;
 
     /** @var Session[] */
-    protected $activeSessions = [];
+    protected array $activeSessions = [];
 
     /**
      * OpenVidu constructor.
-     * @param string $urlOpenViduServer
-     * @param string $secret
-     * @param bool   $sslCheck
+     * @param RestClient $restClient
      */
-    public function __construct(string $urlOpenViduServer, string $secret, bool $sslCheck = true)
+    public function __construct(RestClient $restClient)
     {
-        $this->secret = $secret;
-        $this->sslCheck = $sslCheck;
-        $this->urlOpenViduServer = $urlOpenViduServer;
-        $this->restClient = $this->buildRestClient();
+        $this->restClient = $restClient;
     }
 
     /**
@@ -67,11 +41,12 @@ class OpenVidu
     {
         $session = Session::createFromProperties($this->restClient, $properties);
         $this->activeSessions[$session->getSessionId()] = $session;
+
         return $session;
     }
 
     /**
-     * @param SessionProperties|null $properties
+     * @param mixed[] $data
      * @return Session|null
      */
     public function createSessionFromDataArray(array $data): ?Session
@@ -79,19 +54,11 @@ class OpenVidu
         try {
             $session = Session::createFromArray($this->restClient, $data);
             $this->activeSessions[$session->getSessionId()] = $session;
+
             return $session;
-        }catch (EnumException $e){
+        } catch (InvalidDataException $e) {
             return null;
         }
-    }
-
-    /**
-     * @return RestClient
-     */
-    private function buildRestClient(): RestClient
-    {
-        $client = new RestClient($this->urlOpenViduServer, self::HTTP_USERNAME, $this->secret, $this->sslCheck);
-        return $client;
     }
 
     /**
@@ -102,31 +69,29 @@ class OpenVidu
      */
     public function startRecording(string $sessionId, ?RecordingProperties $properties = null): Recording
     {
-        if (!$properties) {
+        if (null === $properties) {
             $properties = (new RecordingPropertiesBuilder())->build();
         }
         try {
             $data = [
-                "session" => $sessionId,
-                "name" => $properties->getName(),
-                "outputMode" => (string)$properties->getOutputMode(),
+                'session' => $sessionId,
+                'name' => $properties->getName(),
+                'outputMode' => (string)$properties->getOutputMode(),
                 'hasAudio' => $properties->isHasAudio(),
-                'hasVideo' => $properties->isHasVideo()
+                'hasVideo' => $properties->isHasVideo(),
             ];
             if ($properties->getOutputMode()->equalsString(RecordingOutputModeEnum::COMPOSED)) {
                 $data['resolution'] = $properties->getResolution();
-                $data['recordingLayout'] = (string)($properties->getRecordingLayout() ?? "");
+                $data['recordingLayout'] = (string)$properties->getRecordingLayout();
                 if ($properties->getRecordingLayout()->equalsString(RecordingLayoutEnum::CUSTOM)) {
-                    $data['customLayout'] = $properties->getCustomLayout() ?? "";
+                    $data['customLayout'] = $properties->getCustomLayout() ?? '';
                 }
             }
-            $arrayRecording = $this->restClient->post(self::RECORDINGS_START_URL, $data)->getArray();
+            $arrayRecording = $this->restClient->post(ApiPaths::RECORDINGS_START, $data)->getArray();
 
             return new Recording($arrayRecording);
-        } catch (RestClientException $e) {
-            throw new OpenViduException("Could not start recording", 0, $e);
-        } catch (EnumException $e) {
-            throw new OpenViduException("Could not start recording", 0, $e);
+        } catch (RestClientException | InvalidDataException $e) {
+            throw new OpenViduException('Could not start recording', 0, $e);
         }
     }
 
@@ -138,13 +103,12 @@ class OpenVidu
     public function stopRecording(string $recordingId): Recording
     {
         try {
-            $arrayRecording = $this->restClient->post(self::RECORDINGS_STOP_URL . '/' . $recordingId)
+            $arrayRecording = $this->restClient->post(ApiPaths::RECORDINGS_STOP . '/' . $recordingId)
                 ->getArray();
+
             return new Recording($arrayRecording);
-        } catch (RestClientException $e) {
-            throw new OpenViduException("Could not stop recording", 0, $e);
-        } catch (EnumException $e) {
-            throw new OpenViduException("Could not stop recording", 0, $e);
+        } catch (RestClientException | InvalidDataException $e) {
+            throw new OpenViduException('Could not stop recording', 0, $e);
         }
     }
 
@@ -156,12 +120,11 @@ class OpenVidu
     public function getRecording(string $recordingId): Recording
     {
         try {
-            $arrayRecording = $this->restClient->get(self::RECORDINGS_URL . '/' . $recordingId)->getArray();
+            $arrayRecording = $this->restClient->get(ApiPaths::RECORDINGS . '/' . $recordingId)->getArray();
+
             return new Recording($arrayRecording);
-        } catch (RestClientException $e) {
-            throw new OpenViduException("Could not retrieve recording", 0, $e);
-        } catch (EnumException $e) {
-            throw new OpenViduException("Could not retrieve recording", 0, $e);
+        } catch (RestClientException | InvalidDataException $e) {
+            throw new OpenViduException('Could not retrieve recording', 0, $e);
         }
     }
 
@@ -172,17 +135,16 @@ class OpenVidu
     public function listRecordings(): array
     {
         try {
-            $items = $this->restClient->get(self::RECORDINGS_URL)->getArrayInArrayKey('items');
+            $items = $this->restClient->get(ApiPaths::RECORDINGS)->getArrayInArrayKey('items');
 
             $recordings = [];
             foreach ($items as $itemValues) {
                 $recordings[] = new Recording($itemValues);
             }
+
             return $recordings;
-        } catch (RestClientException $e) {
-            throw new OpenViduException("Could not retrieve recordings", 0, $e);
-        } catch (EnumException $e) {
-            throw new OpenViduException("Could not retrieve recordings", 0, $e);
+        } catch (RestClientException | InvalidDataException $e) {
+            throw new OpenViduException('Could not retrieve recordings', 0, $e);
         }
     }
 
@@ -193,9 +155,9 @@ class OpenVidu
     public function deleteRecording(string $recordingId): void
     {
         try {
-            $this->restClient->delete(self::RECORDINGS_URL . '/' . $recordingId);
+            $this->restClient->delete(ApiPaths::RECORDINGS . '/' . $recordingId);
         } catch (RestClientException $e) {
-            throw new OpenViduException("Could not delete recording", 0, $e);
+            throw new OpenViduException('Could not delete recording', 0, $e);
         }
     }
 
@@ -205,7 +167,10 @@ class OpenVidu
      */
     public function getActiveSessions(): array
     {
-        return clone $this->activeSessions;
+        return array_map(
+            fn(Session $session) => clone $session,
+            $this->activeSessions
+        );
     }
 
     /**
@@ -215,7 +180,7 @@ class OpenVidu
     public function fetch(): bool
     {
         try {
-            $jsonArraySessions = $this->restClient->get(self::SESSIONS_URL)->getArrayInArrayKey('content');
+            $jsonArraySessions = $this->restClient->get(ApiPaths::SESSIONS)->getArrayInArrayKey('content');
 
             // Set to store fetched sessionIds and later remove closed sessions
             $fetchedSessionIds = [];
@@ -223,7 +188,7 @@ class OpenVidu
             $hasChanged = false;
 
             foreach ($jsonArraySessions as $arraySession) {
-                $sessionId = $arraySession["sessionId"];
+                $sessionId = $arraySession['sessionId'];
                 $fetchedSessionIds[] = $sessionId;
                 if (isset($this->activeSessions[$sessionId])) {
                     $session = $this->activeSessions[$sessionId];
@@ -238,18 +203,15 @@ class OpenVidu
                 }
             }
             foreach ($this->activeSessions as $key => $session) {
-                if (!in_array($key, $fetchedSessionIds)) {
+                if (!in_array($key, $fetchedSessionIds, true)) {
                     $hasChanged = true;
                     unset($this->activeSessions[$key]);
                 }
             }
+
             return $hasChanged;
-        } catch (RestClientException $e) {
-            throw new OpenViduException('Sessions fetching failed', 0, $e);
-        } catch (EnumException $e) {
+        } catch (RestClientException | InvalidDataException $e) {
             throw new OpenViduException('Sessions fetching failed', 0, $e);
         }
     }
-
-
 }
