@@ -1,34 +1,35 @@
 <?php
 
+declare(strict_types=1);
 
 namespace Stopka\OpenviduPhpClient\Rest;
-
 
 use GuzzleHttp\Exception\RequestException;
 
 class RestResponseException extends RestClientException
 {
-    /** @var RestResponse */
-    private $response;
+    protected const KEY_MESSAGE = 'message';
+
+    /** @var RestResponse|null */
+    private ?RestResponse $response;
 
     public function __construct(RequestException $previous)
     {
-        if ($previous->hasResponse()) {
-            $this->response = new RestResponse($previous->getResponse());
+        $httpResponse = $previous->getResponse();
+        if (null !== $httpResponse) {
+            $this->response = new RestResponse($httpResponse);
+            $message = null;
             try {
-                $message = $this->response->getStringInArrayKey('message');
+                $message = $this->response->getStringInArrayKey(self::KEY_MESSAGE);
             } catch (RestResponseInvalidException $e) {
-
             }
         }
         $message = $message ?? $previous->getMessage();
         parent::__construct($message, $previous->getCode(), $previous);
     }
 
-    public function getResponse(): RestResponse
+    public function getResponse(): ?RestResponse
     {
         return $this->response;
     }
-
-
 }
