@@ -1,46 +1,43 @@
 <?php
-/**
- * Created by IntelliJ IDEA.
- * User: stopka
- * Date: 13.10.17
- * Time: 14:02
- */
+
+declare(strict_types=1);
 
 namespace Stopka\OpenviduPhpClient\Recording;
 
-
-use Stopka\OpenviduPhpClient\Enum;
+use Stopka\OpenviduPhpClient\Enum\EnumTrait;
 
 class RecordingLayoutEnum
 {
-    use Enum;
+    use EnumTrait;
 
     /**
      * All the videos are evenly distributed, taking up as much space as possible
      */
-    const BEST_FIT = "BEST_FIT";
+    public const BEST_FIT = 'BEST_FIT';
 
     /**
      * <i>(not available yet)</i>
      */
-    const PICTURE_IN_PICTURE = "PICTURE_IN_PICTURE";
+    public const PICTURE_IN_PICTURE = 'PICTURE_IN_PICTURE';
 
     /**
      * <i>(not available yet)</i>
      */
-    const VERTICAL_PRESENTATION = "VERTICAL_PRESENTATION";
+    public const VERTICAL_PRESENTATION = 'VERTICAL_PRESENTATION';
 
     /**
      * <i>(not available yet)</i>
      */
-    const HORIZONTAL_PRESENTATION = "HORIZONTAL_PRESENTATION";
+    public const HORIZONTAL_PRESENTATION = 'HORIZONTAL_PRESENTATION';
 
     /**
      * <i>(not available yet)</i>
      */
-    const CUSTOM = "CUSTOM";
+    public const CUSTOM = 'CUSTOM';
 
-
+    /**
+     * @return string[]
+     */
     public function getValues(): array
     {
         return [
@@ -48,7 +45,7 @@ class RecordingLayoutEnum
             self::PICTURE_IN_PICTURE,
             self::VERTICAL_PRESENTATION,
             self::HORIZONTAL_PRESENTATION,
-            self::CUSTOM
+            self::CUSTOM,
         ];
     }
 }
