@@ -1,26 +1,22 @@
 <?php
-/**
- * Created by IntelliJ IDEA.
- * User: stopka
- * Date: 13.10.17
- * Time: 14:02
- */
+
+declare(strict_types=1);
 
 namespace Stopka\OpenviduPhpClient\Recording;
 
-
-use Stopka\OpenviduPhpClient\Enum;
+use Stopka\OpenviduPhpClient\Enum\EnumTrait;
 
 class RecordingModeEnum
 {
-    use Enum;
+    use EnumTrait;
+
     /**
      * The session is recorded automatically as soon as the first client publishes a
      * stream to the session. It is automatically stopped after last user leaves the
      * session (or until you call
      * {@link io.openvidu.java.client.OpenVidu#stopRecording(String)}).
      */
-    const ALWAYS = "ALWAYS";
+    public const ALWAYS = 'ALWAYS';
 
     /**
      * The session is not recorded automatically. To record the session, you must
@@ -28,15 +24,16 @@ class RecordingModeEnum
      * To stop the recording, you must call
      * {@link io.openvidu.java.client.OpenVidu#stopRecording(String)}.
      */
-    const MANUAL = "MANUAL";
+    public const MANUAL = 'MANUAL';
 
+    /**
+     * @return string[]
+     */
     public function getValues(): array
     {
         return [
             self::ALWAYS,
-            self::MANUAL
+            self::MANUAL,
         ];
     }
-
-
 }
