@@ -1,65 +1,56 @@
 <?php
-/**
- * Created by IntelliJ IDEA.
- * User: stopka
- * Date: 22.11.18
- * Time: 16:16
- */
+
+declare(strict_types=1);
 
 namespace Stopka\OpenviduPhpClient\Rest;
 
-use GuzzleHttp\Client;
+use GuzzleHttp\ClientInterface;
 use GuzzleHttp\Exception\RequestException;
 use Psr\Http\Message\ResponseInterface;
 
 class RestClient
 {
-    /** @var Client */
-    private $httpClient;
 
-    const HEADER_ACCEPT = "Accept";
-    const MIME_JSON = "application/json";
+    /** @var ClientInterface */
+    private ClientInterface $httpClient;
+
 
     /**
      * RestClient constructor.
-     * @param string $baseUri
-     * @param string $username
-     * @param string $password
-     * @param bool   $verify
+     * @param ClientInterface $httpClient
      */
-    public function __construct(string $baseUri, string $username, string $password, bool $verify = true)
+    public function __construct(ClientInterface $httpClient)
     {
-        $this->httpClient = new Client([
-            'base_uri' => $baseUri,
-            'auth'     => [$username, $password],
-            'headers'  => [self::HEADER_ACCEPT => self::MIME_JSON],
-            'verify'   => $verify,
-        ]);
-    }
-
-    /**
-     * @param array $data
-     * @return array
-     */
-    private function prepareOptions(array $data = []): array
-    {
-        return [
-            'json' => $data
-        ];
+        $this->httpClient = $httpClient;
     }
 
     /**
      * @param string $url
-     * @param array $data
+     * @param mixed[] $data
      * @return RestResponse
      * @throws RestResponseException
      */
     public function post(string $url, array $data = []): RestResponse
     {
         $options = $this->prepareOptions($data);
-        return $this->processResponse(function () use ($url, $options): ResponseInterface {
-            return $this->httpClient->post($url, $options);
-        });
+
+        return $this->processResponse(
+            fn(): ResponseInterface => $this->httpClient->request(HttpMethodEnum::POST, $url, $options)
+        );
+    }
+
+    /**
+     * @param string $url
+     * @param mixed[] $data
+     * @return RestResponse
+     */
+    public function put(string $url, array $data = []): RestResponse
+    {
+        $options = $this->prepareOptions($data);
+
+        return $this->processResponse(
+            fn(): ResponseInterface => $this->httpClient->request(HttpMethodEnum::PUT, $url, $options)
+        );
     }
 
     /**
@@ -69,9 +60,9 @@ class RestClient
      */
     public function get(string $url): RestResponse
     {
-        return $this->processResponse(function () use ($url): ResponseInterface {
-            return $this->httpClient->get($url);
-        });
+        return $this->processResponse(
+            fn(): ResponseInterface => $this->httpClient->request(HttpMethodEnum::GET, $url)
+        );
     }
 
     /**
@@ -81,9 +72,20 @@ class RestClient
      */
     public function delete(string $url): RestResponse
     {
-        return $this->processResponse(function () use ($url): ResponseInterface {
-            return $this->httpClient->delete($url);
-        });
+        return $this->processResponse(
+            fn(): ResponseInterface => $this->httpClient->request(HttpMethodEnum::DELETE, $url)
+        );
+    }
+
+    /**
+     * @param mixed[] $data
+     * @return mixed[]
+     */
+    private function prepareOptions(array $data = []): array
+    {
+        return [
+            'json' => $data,
+        ];
     }
 
     /**
@@ -94,11 +96,11 @@ class RestClient
     private function processResponse(callable $callback): RestResponse
     {
         try {
-            $response = call_user_func($callback);
+            $response = $callback();
         } catch (RequestException $e) {
             throw new RestResponseException($e);
         }
+
         return new RestResponse($response);
     }
-
 }
