@@ -1,8 +1,8 @@
 <?php
 
+declare(strict_types=1);
 
 namespace Stopka\OpenviduPhpClient\Rest;
-
 
 class RestResponseInvalidException extends RestClientException
 {
